added solution files for project 0x12.

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	unsigned int lnode = 0;
+	size_t lnode = 0;
 
 	if (!h)
 		return (0);
